refactor(CharSet): make it an opaque type instead of an AnyVal value class
Same pattern Subset already uses. An AnyVal value class still boxes
in plenty of realistic contexts - stored in a case class field
(Regex.Chars(set: CharSet)), in a collection, as a type parameter -
whereas an opaque type has zero runtime representation beyond
ArraySeq[Range] itself, in every context, always.

Moved CharSet + Range into their own file (regex/CharSet.scala),
separate from Regex.scala: a top-level opaque type is transparent to
its *whole enclosing file*, not just its companion object. Verified
this empirically before relying on it - a same-file consumer sees
straight through the opacity to ArraySeq/Seq's *native* union/
intersect (list concatenation / multiset intersection, wrong
semantics here) even when calling from outside the type's own
companion object, silently shadowing a colliding extension method
name; a different-file consumer in the same package does not. Since
Regex.alt/Regex.inter call `_ union _`/`_ intersect _` on CharSet
values, keeping CharSet in Regex.scala would have made those silently
wrong. Same reasoning already applies to why Subset lives in its own
file.

Within CharSet.scala itself, `union`/`intersect` are implemented as
plain private functions (unionImpl/intersectImpl) taking explicit
ArraySeq[Range] parameters rather than extension methods calling
themselves by name, sidestepping the exact same collision hazard for
the internal implementation.

No public API or behavior change - drops the @publicInBinary/
private[CharSet]-constructor dance entirely (opaque types don't need
it: CharSet.of is an ordinary public method, safely callable from
macro-generated code in any downstream package without special
annotations).

Full test suite green (CharSetTest 29/29, RegexInterpolatorTest 17/17
exercising the ToExpr/FromExpr macro path), scalafmt clean, compiles
on jvm/scala-js and under --jmh.
