+++ regex/Regex.scala
@@ -2,9 +2,9 @@ package halotukozak.regex
 
 import halotukozak.commons.deepRecursive
 
-import scala.annotation.{publicInBinary, tailrec, threadUnsafe, unused}
-import scala.collection.immutable.{ArraySeq, SortedSet}
-import scala.quoted.{Expr, FromExpr, Quotes, ToExpr, Varargs}
+import scala.annotation.{threadUnsafe, unused}
+import scala.collection.immutable.SortedSet
+import scala.quoted.{Expr, Quotes, ToExpr}
 import scala.util.hashing.MurmurHash3
 
 /**
@@ -248,207 +248,6 @@ object Regex:
         '{ Regex.inter($partsExpr) }
       case Star(inner) => '{ ${ Expr(inner) }.star }
       case Compl(inner) => '{ ! ${ Expr(inner) } }
-  // $COVERAGE-ON$
-
-/**
- * Sorted, non-overlapping, merged ranges over Int code points.
- *
- * The type itself is public - `given ToExpr[Regex]` needs to name it in code spliced into
- * arbitrary external packages by macros - but the primary constructor is restricted to
- * [[CharSet]]'s own companion so every instance is still built through a normalizing smart
- * constructor ([[CharSet.normalize]], [[CharSet.single]], [[CharSet.range]], [[CharSet.empty]],
- * [[CharSet.all]]); `complement` and `intersect` rely on that invariant.
- *
- * Backed by [[ArraySeq]] rather than `Vector`: this type is immutable once built (every
- * instance goes through the normalizing smart constructors above, never incremental
- * append/prepend), so there's nothing to gain from `Vector`'s structural sharing - but
- * `contains`'s binary search, and `intersect`/`complement`'s indexed loops, benefit from
- * `ArraySeq`'s true O(1) array-backed indexed access (`Vector`'s is O(log32 n)) while still
- * getting structural `equals`/`hashCode` for free (unlike a raw `Array`), which `Regex`'s
- * `Set`-based ACI normalization and cached `hashCode` rely on.
- */
-final class CharSet @publicInBinary private[CharSet] (private val ranges: ArraySeq[Range]) extends AnyVal:
-
-  def contains(c: Int): Boolean =
-    @tailrec
-    def loop(lo: Int, hi: Int): Boolean =
-      if lo > hi then false
-      else
-        val mid = (lo + hi) >>> 1
-        val r = ranges(mid)
-        if c < r.lo then loop(lo, mid - 1)
-        else if c > r.hi then loop(mid + 1, hi)
-        else true
-    loop(0, ranges.length - 1)
-
-  def isEmpty: Boolean = ranges.isEmpty
-
-  /**
-   * Two-pointer merge of the two already-sorted, already-coalesced range sequences, O(n+m) -
-   * as opposed to flattening both into one list and re-sorting from scratch (what
-   * `CharSet.normalize(ranges ++ other.ranges)` used to do here, at O((n+m) log(n+m))), which
-   * throws away the fact that both inputs are already ordered. Mirrors `intersect`'s existing
-   * two-pointer walk, just coalescing overlapping/adjacent ranges (`normalize`'s rule) instead
-   * of intersecting them.
-   *
-   * Writes straight into an `ArraySeq.newBuilder`, unlike `intersect`/`complement`/`normalize`
-   * which accumulate into a `Vector` and convert via `ArraySeq.from` at the end - skips both
-   * the `Vector` node allocations and that final copy.
-   */
-  infix def union(other: CharSet): CharSet =
-    val builder = ArraySeq.newBuilder[Range]
-    builder.sizeHint(ranges.length + other.ranges.length)
-    var i = 0
-    var j = 0
-    var cur: Range | Null = null
-    while i < ranges.length || j < other.ranges.length do
-      val takeFromA = j >= other.ranges.length || (i < ranges.length && ranges(i).lo <= other.ranges(j).lo)
-      val next = if takeFromA then ranges(i) else other.ranges(j)
-      if takeFromA then i += 1 else j += 1
-      cur match
-        case null => cur = next
-        case c if next.lo <= c.hi + 1 => cur = Range(c.lo, math.max(c.hi, next.hi))
-        case c =>
-          builder += c
-          cur = next
-    cur match
-      case null => ()
-      case c => builder += c
-    CharSet(builder.result())
-  def |(other: CharSet): CharSet = union(other)
-
-  infix def intersect(other: CharSet): CharSet =
-    // No sizeHint here (unlike union/normalize): the actual intersection is very often much
-    // smaller than min(ranges.length, other.ranges.length) - e.g. near-empty for disjoint or
-    // barely-overlapping sets, a common case (`[a-z] & [A-Z]`) - so hinting that upper bound
-    // pre-allocates array capacity that's usually mostly wasted, which benchmarked as a net
-    // loss versus just letting the builder grow on demand.
-    val builder = ArraySeq.newBuilder[Range]
-    var i = 0
-    var j = 0
-    while i < ranges.length && j < other.ranges.length do
-      val a = ranges(i)
-      val b = other.ranges(j)
-      val lo = math.max(a.lo, b.lo)
-      val hi = math.min(a.hi, b.hi)
-      if lo <= hi then builder += Range(lo, hi)
-      if a.hi < b.hi then i += 1 else j += 1
-    CharSet(builder.result())
-
-  def &(other: CharSet): CharSet = intersect(other)
-
-  def complement: CharSet =
-    // Walks `ranges` by index rather than `.head`/`.tail`: unlike `Vector`, slicing off the
-    // head of an array-backed `ArraySeq` is O(n) (a fresh array copy), which would make this
-    // loop O(n^2) overall instead of O(n).
-    val builder = ArraySeq.newBuilder[Range]
-    var idx = 0
-    var prev = 0
-    while idx < ranges.length do
-      val head = ranges(idx)
-      if prev <= head.lo - 1 then builder += Range(prev, head.lo - 1)
-      prev = head.hi + 1
-      idx += 1
-    if prev <= CharSet.maxCodePoint then builder += Range(prev, CharSet.maxCodePoint)
-    CharSet(builder.result())
-
-  def iterator: Iterator[Range] = ranges.iterator
-
-object CharSet:
-
-  /** Upper bound used for complement. Covers all valid Unicode code points. */
-  val maxCodePoint: Int = 0x10ffff
-
-  val empty: CharSet = CharSet(ArraySeq.empty)
-  val all: CharSet = CharSet(ArraySeq(Range(0, maxCodePoint)))
-
-  /** What `.` matches in Java regex without DOTALL flag — all code points except line terminators. */
-  val dotDefault: CharSet = normalize(
-    Range(0, '\n'.toInt - 1),
-    Range('\n'.toInt + 1, '\r'.toInt - 1),
-    Range('\r'.toInt + 1, 0x84),
-    Range(0x86, 0x2027),
-    Range(0x202a, maxCodePoint),
-  )
-
-  def single(c: Char): CharSet = single(c.toInt)
-  def single(c: Int): CharSet = CharSet(ArraySeq(Range(c, c)))
-  def range(lo: Char, hi: Char): CharSet = range(lo.toInt, hi.toInt)
-  def range(lo: Int, hi: Int): CharSet = CharSet(ArraySeq(Range(lo, hi)))
-
-  /** Builds a normalized [[CharSet]] from arbitrary (possibly overlapping) ranges. */
-  def normalize(rs: Iterable[Range]): CharSet =
-    val sorted = rs.toArray
-    sorted.sortInPlaceBy(_.lo)
-    val builder = ArraySeq.newBuilder[Range]
-    builder.sizeHint(sorted.length)
-    var cur: Range | Null = null
-    var i = 0
-    while i < sorted.length do
-      val r = sorted(i)
-      cur match
-        case null => cur = r
-        case c if r.lo <= c.hi + 1 => cur = Range(c.lo, math.max(c.hi, r.hi))
-        case c =>
-          builder += c
-          cur = r
-      i += 1
-    cur match
-      case null => ()
-      case c => builder += c
-    CharSet(builder.result())
-
-  def normalize(ranges: Range*): CharSet = normalize(ranges)
-
-  // $COVERAGE-OFF$
-  given ToExpr[CharSet]:
-    def apply(s: CharSet)(using Quotes): Expr[CharSet] =
-      val rangeExprs: Seq[Expr[Range]] = s.ranges.map(Expr.apply)
-      '{ CharSet(ArraySeq(${ Varargs(rangeExprs) }*)) }
-
-  given FromExpr[CharSet]:
-    private def fromRanges(exprs: Seq[Expr[Range]])(using Quotes): Option[CharSet] =
-      val builder = ArraySeq.newBuilder[Range]
-      builder.sizeHint(exprs.length)
-      val it = exprs.iterator
-      var ok = true
-      while ok && it.hasNext do
-        it.next() match
-          case Expr(range) => builder += range
-          case _ => ok = false
-      if ok then Some(new CharSet(builder.result())) else None
-
-    override def unapply(s: Expr[CharSet])(using Quotes): Option[CharSet] = s match
-      case '{ CharSet(ArraySeq(${ Varargs(rangeExprs) }*)) } => fromRanges(rangeExprs)
-      case '{ new CharSet(ArraySeq(${ Varargs(rangeExprs) }*)) } => fromRanges(rangeExprs)
-      case '{ CharSet.single(${ Expr(c) }: Char) } => Some(CharSet.single(c))
-      case '{ CharSet.single(${ Expr(c) }: Int) } => Some(CharSet.single(c))
-      case '{ CharSet.range(${ Expr(lo) }: Char, ${ Expr(hi) }: Char) } => Some(CharSet.range(lo, hi))
-      case '{ CharSet.range(${ Expr(lo) }: Int, ${ Expr(hi) }: Int) } => Some(CharSet.range(lo, hi))
-      case _ => None
-
-// $COVERAGE-ON$
-
-/** Single closed code-point range. */
-final case class Range(lo: Int, hi: Int):
-  require(0 <= lo && lo <= hi && hi <= CharSet.maxCodePoint, s"invalid code-point range [$lo, $hi]")
-
-  def this(lo: Char, hi: Char) = this(lo.toInt, hi.toInt)
-
-  def contains(c: Int): Boolean = c >= lo && c <= hi
-object Range:
-  // $COVERAGE-OFF$
-  given ToExpr[Range]:
-    def apply(x: Range)(using Quotes): Expr[Range] =
-      '{ Range(${ Expr(x.lo) }, ${ Expr(x.hi) }) }
-
-  given FromExpr[Range]:
-    override def unapply(x: Expr[Range])(using Quotes): Option[Range] = x match
-      case '{ Range(${ Expr(lo) }: Char, ${ Expr(hi) }: Char) } => Some(Range(lo, hi))
-      case '{ Range(${ Expr(lo) }: Int, ${ Expr(hi) }: Int) } => Some(Range(lo, hi))
-      case '{ new Range(${ Expr(lo) }: Char, ${ Expr(hi) }: Char) } => Some(Range(lo, hi))
-      case '{ new Range(${ Expr(lo) }: Int, ${ Expr(hi) }: Int) } => Some(Range(lo, hi))
-      case _ => None
   // $COVERAGE-ON$
 
 extension [A, CC[X] <: Iterable[X]](xs: scala.collection.IterableOps[A, CC, CC[A]])
